feat(formatter): split into functions
